iangelman #249 handle failure of client side validation

+++ rdrf/registry/groups/models.py
@@ -5,6 +5,7 @@ from django.contrib.auth.models import AbstractBaseUser, UserManager, Permission
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 from django.db import models
+from django.db import transaction
 from django.dispatch import receiver
 
 from registration.signals import user_activated
@@ -14,7 +15,6 @@ from rdrf.models.definition.models import Registry
 from registry.groups import GROUPS as RDRF_GROUPS
 
 import logging
-
 logger = logging.getLogger(__name__)
 
 
@@ -264,12 +264,21 @@ def user_registered_callback(sender, user, request, **kwargs):
         from mtm.patient_registration import MtmRegistration
         patient_reg = MtmRegistration(user, request)
 
-    try:
-        patient_reg.process()
-    except Exception as ex:
-        logger.error("Error during registration of %s (rolling back): %s" % (user.username, ex))
+    cleanup_required = False
+
+    with transaction.atomic():
+        try:
+            patient_reg.process()
+        except Exception as ex:
+            logger.error("Error during registration of %s (rolling back): %s" % (user.username, ex))
+            cleanup_required = True
+
+    if cleanup_required:
+        #  if there are unhandled errors in the registration process
+        # we need to do this outside of the atomic block
+        # see angelman #249
+        logger.info("Error in registration requires orphan user cleanup deletion on user %s" % user.username)
         user.delete()
-
